backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pdfplumber==0.11.0
pypdf==4.2.0
python-docx==1.1.2
python-dotenv==1.0.1
gunicorn==22.0.0
